requirements.txt
psycopg2-binary
python-dotenv
streamlit
sqlalchemy
pandas
requests
python-dotenv
beautifulsoup4
plotly
